video/requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
moviepy>=1.0.3
webrtcvad>=2.0.10
numpy>=1.21.0
click>=8.0.0
tqdm>=4.64.0
colorama>=0.4.6